chore(mjs): rename .bin -> /scripts
affects: @findify/mjs
