+++ packages/mjs/webpack.config.js
@@ -1,2 +1,2 @@
 require('babel-register');
-module.exports = require('./.bin');
+module.exports = require('./scripts');
